New ux for editing title, update in separate component
