+++ src/components/CustomCollection/CustomCollection.tsx
@@ -1,27 +1,23 @@
-import React, { useState, useRef, useEffect, FormEvent } from 'react'
+import React, { useState, useRef, useEffect } from 'react'
 import {
   Button,
-  ButtonGroup,
   Form,
   Label,
   ModalRef,
   ComboBox,
   ComboBoxOption,
   ComboBoxRef,
-  TextInput,
 } from '@trussworks/react-uswds'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-// import { EditableCollectionTitle } from './EditableCollectionTitle'
+import { EditableCollectionTitle } from './EditableCollectionTitle'
 import { RemovableBookmark } from './RemovableBookmark'
 import styles from './CustomCollection.module.scss'
-
 import Collection from 'components/Collection/Collection'
 import type { Bookmark as BookmarkType, BookmarkRecords } from 'types/index'
 import AddCustomLinkModal from 'components/modals/AddCustomLinkModal'
 import DropdownMenu from 'components/DropdownMenu/DropdownMenu'
 import RemoveCustomCollectionModal from 'components/modals/RemoveCustomCollectionModal'
 import { useCloseWhenClickedOutside } from 'hooks/useCloseWhenClickedOutside'
-
 // Not an ideal way to validate URLs but this will work for now
 const VALID_URL_REGEX = /^(ftp|http|https):\/\/[^ "]+$/
 
@@ -35,89 +31,6 @@ type PropTypes = {
   handleRemoveCollection: () => void
   handleEditCollection: (title: string) => void
 }
-
-// Workspace for edit collection title
-
-type TitlePropTypes = {
-  collectionId: string
-  text: string
-  onSave: (event: React.FormEvent<HTMLFormElement>) => void
-  onCancel: () => void
-  onDelete: () => void
-  isEditing: boolean
-}
-
-export const EditableCollectionTitle = ({
-  collectionId,
-  text,
-  isEditing,
-  onSave,
-  onCancel,
-}: TitlePropTypes) => {
-  // const [isEditing, setEditing] = useState(isActive)
-  const [currentText, setCurrentText] = useState(text)
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  const handleCancel = () => {
-    console.log('handleCancel')
-    const inputEl = inputRef.current as HTMLInputElement
-    inputEl.value = ''
-    onCancel()
-    // TODO reset form and set active to false, move this to parent
-  }
-  useEffect(() => {
-    if (isEditing) {
-      inputRef?.current?.focus()
-    }
-  }, [isEditing])
-
-  const inputId = `collectionTitle_${collectionId}`
-
-  return (
-    <>
-      {isEditing ? (
-        <Form onSubmit={onSave}>
-          <Label htmlFor={inputId} className="usa-sr-only">
-            Collection Title
-          </Label>
-          <TextInput
-            inputRef={inputRef}
-            id={inputId}
-            name="collectionTitle"
-            required
-            maxLength={200}
-            className={styles.collectionTitle}
-            type="text"
-            // value={currentText}
-            // placeholder="Add a title for this collection"
-          />
-          <ButtonGroup>
-            <Button type="submit" data-close-modal>
-              Save link name
-            </Button>
-            <Button
-              type="button"
-              data-close-modal
-              unstyled
-              className="padding-105 text-center"
-              onClick={handleCancel}>
-              Cancel
-            </Button>
-          </ButtonGroup>
-        </Form>
-      ) : (
-        <h3
-          // tabIndex={0}
-          role="button"
-          className={styles.collectionTitle}>
-          {currentText}
-        </h3>
-      )}
-    </>
-  )
-}
-
-// end workspace for edit collection title
 
 const CustomCollection = ({
   _id,
@@ -134,9 +47,14 @@ const CustomCollection = ({
   const addCustomLinkModal = useRef<ModalRef>(null)
   const linkInput = useRef<ComboBoxRef>(null)
   const [isEditing, setEditing] = useState(false)
+
   useEffect(() => {
     if (isAdding && linkInput.current) {
       linkInput.current.focus()
+    }
+    // If there is no title, prompt user to enter one
+    if (title === '') {
+      setEditing(true)
     }
   }, [isAdding])
 
@@ -144,8 +62,8 @@ const CustomCollection = ({
     e.preventDefault()
     const data = new FormData(e.currentTarget)
     const label = `${data.get('collectionTitle')}`
-    // resetForm()
     handleEditCollection(label)
+    setEditing(false)
   }
 
   const handleShowAdding = () => setIsAdding(true)
